bugfix: fix beeps on GTK field overflow
If the calculations are just right and the `snprintf` calls used to populate
the GTK fields exceed the `max-length` field defined in resources/xnec2c.glade,
then GTK will send a system beep event to the user indicating an invalid input.

However, since it wasn't really the user's input, but rather a bug that
overflowed a field, this bug was very difficult to find and troubleshoot.  Now
we know if someone gets continuous beeping while rotating an object, changing a
graph or resizing a window, then there is most likely a GTK text field being
set to a value beyond its defined maximum length.

(This beeping issue was reported by a user long ago, but I cannot find the
reference.  So, to the original bug reporter, thank you for reporting the
issue.  I'm glad we were able to find a reproducer, and now it should be solved.)