prevent hang when freq plot window is too small
Thanks to Tom DL1JBE who reports:

    For any NEC design if I open the Frequency Plot window in its
    default size and select more than 4 of the Diagrams (Max Gain,
    Direction, ...) xnec2c stops responding.  But if I first resize the
    frequency plot window to a much bigger size I can select more than 5
    buttons at the same time. Resizing it to a smaller size again it
    will hang up again.

This was caused by the vertical scale calculator: when the vertical
scale is small enough it hangs while calculating the vertical
subdivisions.  The fix is that if there are fewer than 3 subdivisions,
we just don't do the calculation.

Relatedly, we solved the problem where small freqplot windows would not
show min/max values, now they do.