glade: fix GTK crash related to menuitem icon
There seems to be a problem with GTK's Glade XML builder that causes a
crash when an icon alias is created and then re-used multiple times.
This commit fixes the issue reported here when building for flat pack:

	https://github.com/flathub/org.xnec2c.Xnec2c/pull/8

Thank you to @AsciiWolf for reporting this and helping me troubleshoot.

Also see this question on stack exchange. If amount knows how to answer
the question, I think it solves the the issue:

	https://stackoverflow.com/questions/73603162/why-does-gtk-builder-add-objects-tell-me-object-with-id-saveas-csv-not-found