add config checkbox: Begin Frequency Loop at Startup
Previously, the frequency loop would auto-start when xnec2c loaded if
the play button had been pressed to save that setting in a previous run
of xnec2c. This was not a particularly intuitive feature, and it has not
been clear to users when or why the frequency loop will start at program
load. This commit adds a menu item to explicitly enable or disable
running the frequency loop at startup. Additionally, the frequency loop
is always enabled if batch mode (--batch) or optimization mode
(--optimize) is enabled.

Thanks to @ProtonTheFox for help in troubleshooting issue #44 and #45.

Fixes: #45