Prevent deadlock when Notice() is called
It is possible that we reach the Notice function while a lock was held.

Because GTK will iterate queued events during `gtk_dialog_run`, GTK may call a
function that needs the lock, causing a deadlock. Avoid the deadlock by
printing the Notice only to the terminal as an error. This is an unlikely edge
case and if this gets hit, then the user is probably using the terminal to
begin with for some type of batch processing or program testing.