Prevent deadlock on Open_Input_File
If there is an error in the input .NEC file while the file is being
opened and a new frequency data set is being calculated, there is a
deadlock between retrying the currents (and charges) while the file is
still being opened.

Move the redraw to the bottom of the open function after the lock has
been released to prevent that.