add GTK menu tear-offs
Certain menus have many options and it can be tedious to select the menu to
click commonly used options one after the next. This adds a tear-off handle to the
top of the menu so that it will open the menu as a new window.

Officially this tear-off feature is deprecated in GTK 3.4, but it is very
useful and has not been removed. This commit may need to be reverted someday,
but it should be stable at least until 2034 when extended support for Ubuntu
24.04 expires.