make pan/tilt/zoom smooth during optimization
Optimization used to suspend drawing while the frequency loop is running.
This is good because it prevents flicker, but it causes poor user interaction
because of the delay between iterations for slower optimizations.  This
commit adds an exception to the redraw processing so that UI actions may
trigger a redraw